2.5.2 release

+++ play.it-2/lib/libplayit2.sh
@@ -32,8 +32,8 @@
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-library_version=2.5.1
-library_revision=20180211.4
+library_version=2.5.2
+library_revision=20180219.2
 
 # set package distribution-specific architecture
 # USAGE: set_architecture $pkg
@@ -1549,10 +1549,16 @@ write_bin() {
 			    local file_real
 			    cd "$1"
 			    find . -type f | while read -r file; do
-			      file_prefix="$(readlink -e "$PATH_PREFIX/$file")"
+			      if [ -e "$PATH_PREFIX/$file" ]; then
+			        file_prefix="$(readlink -e "$PATH_PREFIX/$file")"
+			      else
+			        unset file_prefix
+			      fi
 			      file_real="$(readlink -e "$file")"
 			      if [ "$file_real" != "$file_prefix" ]; then
-			        rm --force "$PATH_PREFIX/$file"
+			        if [ "$file_prefix" ]; then
+			          rm --force "$PATH_PREFIX/$file"
+			        fi
 			        mkdir --parents "$PATH_PREFIX/${file%/*}"
 			        ln --symbolic "$file_real" "$PATH_PREFIX/$file"
 			      fi
